Initial SQLite3 (v3.2.2) port.
Took a bit to get this committed due a bunch of other projects under way...
However, here it finally is (and bound to be buggy ;-) of course).  There's yet
no library - only the executable (sqlite).  I'll throw in the library code
gradually, when having free time from other chores.

NOTE: I've not tested *everything* with this (as AROS lacks Tcl - anyone want
to port it?) - so - feedback about buggy operations and such is *greatly
appreciated*!


git-svn-id: https://svn.aros.org/svn/aros/trunk@23464 fb15a70f-31f2-0310-bbcc-cdcc74a49acc