Build cybergraphics.library.....
This one was a bit dirty hack. The cybergraphics.library in AROS does not replace graphics.library - it extends it. Therefore both use the same graphics driver. In order to make cybergraphics work, I make it as a kobj linkable module, and link with graphics.library in order to create a packet of two :)

git-svn-id: https://svn.aros.org/svn/aros/trunk@26830 fb15a70f-31f2-0310-bbcc-cdcc74a49acc