rom/dos: Handle filesystems that continue to return ST_USERDIR instead of ST_ROOT
The UAE filesystem ROM continues to return ST_USERDIR for '/',
even at the root directory. According to Toni, several 'classic'
AmigaOS Packet filesystems do this too.

This patch detects this condition by checking to see if
io_DirPos (aka fh_DiskKey) is the same on a second invocation
of FSA_EXAMINE. If so, the ST_USERDIR is assumed to be (in reality)
a ST_ROOT.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk@35613 fb15a70f-31f2-0310-bbcc-cdcc74a49acc