Overhaul the way how GL context is store - no longer use fields in libbase and a global registry variable to store the context.
The reasons for this change are:
a) the GL context is actually not "current" per library opener but per task
b) while the global variable was reserving the registry in mesa, any other
called code could clober the registry and lead to crashes

The new approach use naive TLS implementation for all contexts: GL, EGL and
OpenVG. This implementation is slower than the registry based mainly due
to need to synchronize on semaphore. The speed can be made comparable with
registry based solution if context reads are made in a way that they don't
require synchonization on semaphore.

git-svn-id: https://svn.aros.org/svn/aros/trunk@36473 fb15a70f-31f2-0310-bbcc-cdcc74a49acc