graphics.library: Fix (general case) bitter management
Without an interrupt handler for the Blitter that handled the
GfxBase->BlitWaitQ dispatching, the old code would hang when there
was blitter contention.

For the general case, where there is no Blitter interrupt handler, use
the existing blitter semaphore to handle blitter arbitration.

One day, for m68k, we'll need to implement a Blitter interrupt handler
to properly support QBSBlit and QBlit, but that's for another day.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk@38879 fb15a70f-31f2-0310-bbcc-cdcc74a49acc