Report input focus when window is active and cursor is within inner window area
Previously it was enough for window to be active to report input focus.
This was making it impossible to move window when in mouse grab mode.
Now, the input focus is only reported if window is active AND mouse
enters the inner area of window. This allows dragging the window by title
bar or resizing using window gadget.

git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@39167 fb15a70f-31f2-0310-bbcc-cdcc74a49acc