arossupport: Implement alternative global stack that can be used without interference with function argument passing.
Implementation for i386, x86_64 and m68k; PPC is untested; ARM has #error in newstackswap.s

Add initialization so that programs started with NewAddTask or programs
that use NewStackSwap may use these functions. AddTask and StackSwap are
not supported and are deprecated.

Stack location from kickstart is passed into PrepareExecBase so it can be
used for the boot task.

This patch is co-developed with Jason McMullan.

git-svn-id: https://svn.aros.org/svn/aros/trunk@41077 fb15a70f-31f2-0310-bbcc-cdcc74a49acc