Introduce AROSMesaGetProcAddressInternal
This is a private function which resides inside mesa.library and can be
called to check if a certain GL extension function is present in current
3D subsystem configuration. The need for this function was triggered by
work on HostGL implementation where host drivers support different
subset of GL functions than Mesa.

This is ABIV0 solution. For ABIV1 the implementation of
AROSMesaGetProcAddress should be revisited when migrating to C calling
convention.

git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@41116 fb15a70f-31f2-0310-bbcc-cdcc74a49acc