dos.library: Make ReadItem() bug compatible (ugh) with AROS.
AOS ReadItem() is a nasty piece of work.

Unfortunately, a number of applications from early AOS expect it to behave in
the strange ways that it does, and even up to AOS 3.x, these quirks are
preserved.

For those porting applications from AOS, we should be as bug-compatible
with AOS as possible. New applications should not use this routine!

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk@41155 fb15a70f-31f2-0310-bbcc-cdcc74a49acc