tools/genmodule: Fix error in tests/clib/execl; should also fix gcc
This bug was quite complicated.
The trigger was that __exec_do() uses DOS function RunCommand() so the child runs in the same Task structure as the parent.
Now the child can call arosc.library functions overwriting the task memory slot holding the libbase.
When __exec_do then called _exit() after the RunCommand() it would jump into the startup code of the now defunct program.

This patch only fixes the problem for pertaskbase libraries and not for peropenerbase libraries.
A more involved refactoring patch will come later that should fix it for all cases.

git-svn-id: https://svn.aros.org/svn/aros/trunk@43841 fb15a70f-31f2-0310-bbcc-cdcc74a49acc