exec.library - introduce supervisor guru task
If there is non AT_DeadEnd Alert thrown when in supervisor mode,
the system will first delegate showing it to this task. If this task
fails, the regular Exec_SystemAlert path will be continued.

Note that unlike traps where default handler make the task suspend
in generic way, each time a non AT_DeadEnd Alert is fired at
supervisor level it will be systems responsibility to contain
the situation at least long enough for user to actually
see the requester.

git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@46149 fb15a70f-31f2-0310-bbcc-cdcc74a49acc