cdfs-handler: ISO9660 CDROM filesystem
Intended for use where CDVDFS is not licensed.

Will be extended with RockRidge + Amiga support in
the near future.

The infrastruture is a bit more flexible than needed,
so that adding HFS and UDF support should be easy.

Also, filesystem designers may want to re-use the
'bcache.c' and 'bcache.h' block cache subsystems,
which implements a LRU + pseudo-random replacement
block cache.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk@46452 fb15a70f-31f2-0310-bbcc-cdcc74a49acc