graphics.hidd: introduce optimized color conversion routines to replace generic true_to_true function
true_to_true was profiled to use 18% of time when moving windows on
Wanderer. Optimized functions use up to 6% when doing the same.

Functions code taken from patchrgbconv. The patch was use for more than
a year in nouveau driver and now is included into generic code so that
all drivers can benefit from it.

git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@46946 fb15a70f-31f2-0310-bbcc-cdcc74a49acc