country changes ---------------
Rework makecountry, and country source files, so that countries have embedded version string the same as on AmigaOS.  Use v44.0 as the default (with today's date) unless the file already specified a version.  Clean-up the countries source files so that they are consistent.  Embed used flag in the generated country file and remove duplicate flags where appropriate (renamed duplicates to use English name).

git-svn-id: https://svn.aros.org/svn/aros/trunk@47214 fb15a70f-31f2-0310-bbcc-cdcc74a49acc