Backport of r47325:
Added an empty BeginRefresh()/EndRefresh() pair, to ensure that the
console window's damage list is cleared after a resize (we probably
shouldn't use WFLG_NOCAREREFRESH, because the window doesn't belong
to console.device).

This fixes the lock-up that occurred when ClickToFront changes the depth
of a shell window that is being resized or had previously been resized.


git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@47363 fb15a70f-31f2-0310-bbcc-cdcc74a49acc