[csplit] Copied C and math source files to compiler/stdc or compiler/posixc
This is a preparation patch. In next patches functions will be moved from
arosc.library & libm.a to stdc.library, stdcio.library & posixc.library.
If extra changes to the file was needed one will see it as a diff to the
copied source code; original file will be removed in the patch.
The math functions from libm.a have also been copied as they will become part
of shared stdc.library.

Some files have been renamed:
 * __arosc_environ.c -> posixc/__posixc_environ.c
 * __arosc_gmtoffset.c -> stdc/__stdc_gmtoffset.c
 * __arosc_ioerr2errno.c -> stdc/__stdc_ioerr2errno.c
 * __arosc_nixmain.c -> posixc/__posixc_nixmain.c
 * __arosc_startup.c -> stdc/__stdc_startup.c
 * __arosc_termios.h -> posixc/__termios.h
 * __assert.c -> stdc/__stdc_assert.c, stdc/__stdcio_assert.c
   and posixc/__posixc_assert.c