Fixed clean metatargets.
1) build_linklib didn't delete the linklib.

2) fixed double-colon metatargets. I had a test case
with both a linklib and a program. An attempt to clean
the linklib cleaned the program, i.e. the effect was as
if the double-colon target was at the end of the 
Makefile.
Alltough the "make" manual states: "When a target appears
in multiple rules, all the rules must be the same type:
all ordinary, or all double-colon.", using a combination
of an ordinary and a double-colon seems to be the right
workaround.


git-svn-id: https://svn.aros.org/svn/aros/trunk@48339 fb15a70f-31f2-0310-bbcc-cdcc74a49acc