Forward compatibility: genmodule - add option to hide certain function
Option .hidden causes the generated headers to hide the function within
a special define that needs to be explicitly passed by developer.

Functions applicable to be marked as .hidden are functions which are
used by other modules but are either private to that module or are
new functions which have a good chance of being refined in future,
meaning their signature would change.

Those functions are enabled by default when building from build system.
The rationale is that everything that is build from build system:
a) is delivered in nightly build
b) can be rebuild when the function changes