Port from ABIv0:
Converted SMBFS to a normal filesystem handler. I.e. it's no longer the 
program C:SMBFS, but rather L:smb-handler and is started through mount 
files placed in Storage/DOSDrivers.

Additionally, a bug was fixed that allowed directories to be opened as 
files. This prevented the shell's implicit CD from working (i.e. where 
just a directory name is given as a command).


git-svn-id: https://svn.aros.org/svn/aros/trunk@48897 fb15a70f-31f2-0310-bbcc-cdcc74a49acc