NListtree.mcc: fix sorted insert
Since the itaration starts from first element, there is no other
way to insert at head than explicitly return it. This reverts
one of the changes that was done in rev 536  of NList.

Additionally, the built in Head and Tail compare functions had
their implementation switched. The return needs to be < 0 to
iterate further (eventually reaching last element) and >= to
insert at this location, which will be HEAD for first call due
to first change.

The change has been tested with all 5 built in compare functions
and with external compare function.

Change is pushed upstream to NList team.