Fix clock setup on Sam460.
Revert some earlier tweaks to this code that did not fix it and turned
out to be wrong. Add a FIXME comment instead to note that some boards
have a 55MHz sys clock while others have 50MHz. Set it to 50MHz for
now matching QEMU. This could be a problem for some boards but nobody
seems to be using it on real hardware now. If anyone has a board with
55MHz sys clock this FIXME should be fixed but I don't know how.

git-svn-id: https://svn.aros.org/svn/aros/trunk@55081 fb15a70f-31f2-0310-bbcc-cdcc74a49acc